requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dash==1.9.1
dash-auth==1.3.2
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
patsy==0.5.1
plotly==4.5.2
plotly-express==0.4.1
psycopg2-binary==2.8.4
pyorbital==1.5.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
statsmodels==0.11.1
tweepy==3.8.0
ua-parser==0.10.0
urllib3==1.25.8
Werkzeug==1.0.0
